requirements.txt
numpy==2.2.5
packaging==25.0
pandas==2.2.3
pillow==11.2.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
tzdata==2025.2
openpyxl==3.1.5